First working version of Automaton Plotter
We first use GraphViz to get a circular layout and export an SVG file, which then we tweak to obtain a better result (edge labels aligned with the edges, non-obstructing self-loop edges).
Some heuristics tweaking might still be necessary, but this is a good starting point.
Moving on to the trajectory plotter for now.

[ISSUE] unable to extract automaton name
[ISSUE] unable to extract variables' names in dynamic funtions
[ISSUE] unable to extract algebraic functions
[ISSUE] unable to extract invariant functions
[TWEAK] node/edge label content
[TWEAK] heuristics here and there in the svg tweaker