Final bugfixes
Thanks to the latest bindings, the Dashboard is completely independent from the system we choose, even though some options are still not available (i.e. evolver customization, initial conditions).

When plotting 2D axes, the projections are "rigorous", without any additional manipulations. When plotting 3D axes, we project each variable against time and then plot the barycenter of each polytope. This happens because unfortunately there is no way of projection onto N-D planes natively in Ariadne.

Nodes in the Automaton Plotter are now circular and the labels are a little above their edges.

[BUG]   when plotting in the dashboard, an exit(245) occurs without  any specific reason; trajectory_plotter.py does work though
[IDEA]  add a button to import the YAML dump of an orbit
[IDEA]  add more infos of location/transition on the automaton graph
[ISSUE] evolver configuration is not customizable